zap-container-management-python-flask-main/requirements.txt
docker
gunicorn
flask
python-dotenv
python
python3-flask